+++ src/cone/app/tests/__init__.py
@@ -8,6 +8,7 @@ def test_suite():
     from cone.app.tests import test_app
     from cone.app.tests import test_model
     from cone.app.tests import test_security
+    from cone.app.tests import test_ugm
     from cone.app.tests import test_utils
     from cone.app.tests import test_workflow
 
@@ -38,6 +39,7 @@ def test_suite():
     suite.addTest(unittest.findTestCases(test_app))
     suite.addTest(unittest.findTestCases(test_model))
     suite.addTest(unittest.findTestCases(test_security))
+    suite.addTest(unittest.findTestCases(test_ugm))
     suite.addTest(unittest.findTestCases(test_utils))
     suite.addTest(unittest.findTestCases(test_workflow))